renaming files aer_executor -> sim_executor
