+++ test/unit/sim_executor/test_insert_noise_pass.py
@@ -27,7 +27,7 @@ from ...ibm_test_case import IBMTestCase
 if optionals.HAS_AER:
     from qiskit_aer import AerSimulator
 
-    from qiskit_ibm_runtime.aer_executor.insert_noise_pass import InsertNoisePass
+    from qiskit_ibm_runtime.sim_executor.insert_noise_pass import InsertNoisePass
 
 
 def _circuit_with_barrier(n_qubits: int, label: str) -> QuantumCircuit:
