+++ test/unit/sim_executor/test_noisy.py
@@ -31,7 +31,7 @@ from ...ibm_test_case import IBMTestCase
 if optionals.HAS_AER:
     from qiskit_aer import AerSimulator
 
-    from qiskit_ibm_runtime.aer_executor import AerExecutor
+    from qiskit_ibm_runtime.sim_executor import AerExecutor
 
 
 def _batched(iterable, n, *, strict=False):
