+++ test/unit/sim_executor/test_run_quantum_program.py
@@ -26,7 +26,7 @@ from ...ibm_test_case import IBMTestCase
 if optionals.HAS_AER:
     from qiskit_aer import AerSimulator
 
-    from qiskit_ibm_runtime.aer_executor.run_quantum_program import run_quantum_program
+    from qiskit_ibm_runtime.sim_executor.run_quantum_program import run_quantum_program
 
 
 @skipUnless(condition=optionals.HAS_AER, reason="qiskit-aer is required to run this test")
